Ensure that keystrokes sent to desktop chrome AURA don't make their way into accelerator handling in ASH.
ASH registers a focus manager factory which creates focus managers for top level widgets with a FocusManager::Delegate
interface implementation which is used to process accelerators.

We allow only one focus manager factory to exist at any given time, which means that the focus manager factory for
ASH gets notified for subsequent focus manager creation requests, some of them from the desktop. 

This results in keystrokes for desktop widgets making their way to the ASH focus manager delegate. 

Fix is to pass in a bool flag desktop_widget in the NativeWidgetDelegate::OnNativeWidgetCreated method which is implemented
by the Widget class. This flag when true causes the ASH focus manager to default to the default focus manager behavior, which
is to create focus managers with a NULL delegate.

BUG=246821
R=sky@chromium.org

Review URL: https://codereview.chromium.org/15882009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@204331 0039d316-1c4b-4281-b951-d872f2087c98